Backend-flask/requirements.txt
Flask
tensorflow
numpy
Pillow
twilio